# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API for ERPHeydude",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "mauriciopalacio1990@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://rocky-gorge-03730.herokuapp.com/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/generic/getNextNoteIndex/": {
            "get": {
                "tags": [
                    "GENERICAS"
                ],
                "summary": "Obtener ultimo noteindex",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\GeneriController::getNextNoteIndex",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/MC40200/getCurrencyById/{id}": {
            "get": {
                "tags": [
                    "MONEDA"
                ],
                "summary": "Mostrar moneda por id.",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\MC40200Controller::getCurrencyById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Divisa Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/MC40200/addCurrency/{MC40200}": {
            "post": {
                "tags": [
                    "MONEDA"
                ],
                "summary": "Agregar moneda.",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\MC40200Controller::addCurrency",
                "requestBody": {
                    "description": "Crear moneda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MC40200"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/MC40200/updateCurrency/{MC40200}": {
            "put": {
                "tags": [
                    "MONEDA"
                ],
                "summary": "Modificar moneda.",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\MC40200Controller::updateCurrency",
                "requestBody": {
                    "description": "Modificar moneda",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MC40200"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/MC40200/deleteCurrency/{id}": {
            "delete": {
                "tags": [
                    "MONEDA"
                ],
                "summary": "Eliminar moneda.",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\MC40200Controller::deleteCurrency",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Divisa Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/MC40200/getCurrencyAllByLimit/{init}/{limit}": {
            "get": {
                "tags": [
                    "MONEDA"
                ],
                "summary": "Mostrar moneda por id y rango.",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\MC40200Controller::getCurrencyAllByLimit",
                "parameters": [
                    {
                        "name": "init",
                        "in": "path",
                        "description": "Inicio",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Fin",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/SY01400/login/": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Validar usuario",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\SY01400Controller::login",
                "requestBody": {
                    "description": "Validar usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MC40200"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/SY01400/menu": {
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Mostrar menu lateral.",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\SY01400Controller::menu_side",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/SY01400/cards": {
            "post": {
                "tags": [
                    "Menu"
                ],
                "summary": "Mostrar menu fichas.",
                "description": "Returns project data",
                "operationId": "App\\Http\\Controllers\\SY01400Controller::cards",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MC40200"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "MC40200": {
                "title": "MC40200",
                "description": "Tabla para divisas",
                "properties": {
                    "CURNCYID": {
                        "title": "CURNCYID",
                        "description": "Id de moneda",
                        "type": "integer",
                        "format": "int64",
                        "example": 100
                    },
                    "CURRNIDX": {
                        "title": "CURRNIDX",
                        "description": "Indice de moneda",
                        "type": "integer",
                        "format": "int64",
                        "example": 200
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MC40200"
                }
            },
            "MC60100": {
                "title": "MC60100",
                "description": "Tabla para moneda multifuncional",
                "xml": {
                    "name": "MC60100"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "API LARAVEL",
            "description": "API Endpoints of Projects"
        }
    ]
}